css/css-scroll-snap/parsing/scroll-margin-block-end-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-margin-block-end with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-margin-block-end">
<meta name="assert" content="scroll-margin-block-end supports the full grammar '<length>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("scroll-margin-block-end", "-10px");
test_valid_value("scroll-margin-block-end", "calc(2em + 3ex)");
</script>
</body>
</html>
